validaciones

+++ app/src/main/java/com/example/proyecto2/pregunta.java
@@ -24,8 +24,8 @@ public class pregunta {
         }
     }
 
-    public double getRespuesta() {
-        double respuesta = 0;
+    public int getRespuesta() {
+        int respuesta = 0;
         switch (operador) {
             case "+":
                 respuesta = A + B;
